streamprocess/requirements.txt
boto3==1.24.1
opencv-python-headless==4.11.0.86
pinecone-client==6.0.1
deepface==0.0.93
deep-sort-realtime==1.3.2
numpy==2.0.2
python-dotenv==1.0.1
requests==2.32.3
